+++ src/components/LinksPanel/components/LinkElement/LinkElement.styles.ts
@@ -38,5 +38,17 @@ export const getStyles = (theme: GrafanaTheme2, { dynamicFontSize }: { dynamicFo
         background: ${theme.colors.background.secondary};
       }
     `,
+    alignLinkContentLeft: css`
+      display: flex;
+      justify-content: flex-start;
+    `,
+    alignLinkContentCenter: css`
+      display: flex;
+      justify-content: center;
+    `,
+    alignLinkContentRight: css`
+      display: flex;
+      justify-content: flex-end;
+    `,
   };
 };