+++ src/components/LinksPanel/components/LinkElement/LinkElement.tsx
@@ -84,6 +84,15 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
     return () => resize.disconnect();
   }, [dynamicFontSize, linkEl]);
 
+  /**
+   * Mapping of content alignment positions to their corresponding styles.
+   */
+  const alignClassMap = {
+    center: styles.alignLinkContentCenter,
+    left: styles.alignLinkContentLeft,
+    right: styles.alignLinkContentRight,
+  } as const;
+
   if (link.links.length > 1) {
     /**
      * Menu links
@@ -106,7 +115,11 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
       return (
         <Button
           variant="secondary"
-          className={cx(styles.link, gridMode && styles.linkGridMode)}
+          className={cx(
+            styles.link,
+            gridMode && styles.linkGridMode,
+            link.alignContentPosition && alignClassMap[link.alignContentPosition]
+          )}
           size={buttonSize}
           icon={link.icon}
           fill="outline"
@@ -154,14 +167,17 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
 
     if (currentLink.url) {
       return (
-        <div ref={btnRef}>
+        <div
+          ref={btnRef}
+          style={dynamicFontSize ? ({ '--btn-width': `${linkWidth}px` } as React.CSSProperties) : undefined}
+        >
           <LinkButton
             key={currentLink.url}
             className={cx(
               currentLink.isCurrentLink ? styles.currentDashboard : styles.link,
-              gridMode && styles.linkGridMode
+              gridMode && styles.linkGridMode,
+              currentLink.alignContentPosition && alignClassMap[currentLink.alignContentPosition]
             )}
-            style={dynamicFontSize ? ({ '--btn-width': `${linkWidth}px` } as React.CSSProperties) : undefined}
             icon={currentLink.icon}
             href={currentLink.url}
             title={currentLink.name}
@@ -188,7 +204,11 @@ export const LinkElement: React.FC<Props> = ({ link, buttonSize, gridMode = fals
     >
       <Button
         variant="secondary"
-        className={cx(styles.link, gridMode && styles.linkGridMode)}
+        className={cx(
+          styles.link,
+          gridMode && styles.linkGridMode,
+          link.alignContentPosition && alignClassMap[link.alignContentPosition]
+        )}
         key={link.name}
         fill="outline"
         size={buttonSize}