+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the **Business Links** panel are documented in this file. This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added links align customization and current dashboard selection ([#34](https://github.com/VolkovLabs/business-links/issues/34)).
+- Added dynamic font size ([#38](https://github.com/VolkovLabs/business-links/issues/38)).
+
 ## [1.3.0] - 2025-05-28
 
 ### Added